+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2023-09-13
+### Changed
+- **BC** Use CacheTagChecksum to invalidate cache tags. See https://github.com/wieni/wmcontroller_redis/pull/4
+  - Upgrade steps:
+    - Add `wmcontroller.cache.invalidator: wmcontroller.redis.checksum` to your `services.yml` file.
+    - Run `drush wmcontroller_redis:mark-expired` nightly to search for stale entries and mark them for deletion
+
 ## [1.1.0] - 2023-09-13
 ### Added
 - Add Drupal 10 compatibility